Update galga.md
Spreading out format indicators to see if that fixes this one issue.

I noticed that the entire line is already bolded, then I attempt to *rebold* the definition.  I've fixed this in a way that should fix the issue here.  Still means the file didn't fail gracefully, but hopefully it will at least make sense when looking for future fixes.