Fix mistake in description of bump
The bullet points were in the wrong position. Step 1 says to move the sprite backward, that's changing the x and y position. Step 2 says to stop the sprite, that's setting vx and vy to 0. [As written, vx and vy are being set to 0 first, so changing x and y by -1*vx and -1*vy would be changing them by 0, which is nonsensical.]