Arcade Theme Fixes (#6756)
The most notable change here is setting the default arcade theme to legacy instead of light. It also adds the neutral base color to all themes and includes some fixes to dark theme.

Somewhat counter-intuitively, I've updated the Semantic UI blue color as well. This affects our notification popups. Normally I'd switch these to use the theme's colors, but one of the notifications is "Sign-in successful", and signing in also triggers the theme to change based on the user's preferences. The result is that the notification switches colors halfway through, which looks odd. So for now, I decided to just let this fallback on the non-themed color instead, but it required a color update to be easily readable.